refactor: PurchaseFrequencyChart 분리
