+++ apps/frontend/src/components/PurchaseFrequency.tsx
@@ -1,27 +1,14 @@
 import { useState } from 'react'
-import { BarChart, Bar, XAxis, YAxis, CartesianGrid, Tooltip, Legend, ResponsiveContainer } from 'recharts'
-import { usePurchaseFrequencyData } from '../hooks/usePurchaseFrequencyData'
 import styled from '@emotion/styled'
 
 import PurchaseFrequencyFilter from './PurchaseFrequencyFilter'
-const legendFormatter = (key: string) => {
-  return key === 'count' ? '구매횟수' : key
-}
+import PurchaseFrequencyChart from './PurchaseFrequencyChart'
 
-const tooltipFormatter = (value: number, key: string) => {
-  if (key === 'count') {
-    return [`${value}건`, '구매횟수']
-  }
-  return [value, key]
-}
-
-const PurchaseFrequencyChart = () => {
+const PurchaseFrequency = () => {
   const defaultFromDate = '2024-07-01'
   const defaultToDate = '2024-07-31'
   const [fromDate, setFromDate] = useState<string>(defaultFromDate)
   const [toDate, setToDate] = useState<string>(defaultToDate)
-
-  const { data, error, isLoading, isError } = usePurchaseFrequencyData(fromDate, toDate)
 
   return (
     <div>
@@ -35,21 +22,7 @@ const PurchaseFrequencyChart = () => {
         onChangeToDate={setToDate}
       />
 
-      {isLoading && <p>Loading data...</p>}
-      {isError && <p>Error loading data: {error.message}</p>}
-
-      {data && (
-        <ResponsiveContainer width="100%" height={400}>
-          <BarChart data={data} margin={{ top: 20, right: 30, left: 20, bottom: 5 }}>
-            <CartesianGrid strokeDasharray="3 3" />
-            <XAxis dataKey="range" tickFormatter={(value) => value + '원'} />
-            <YAxis />
-            <Tooltip formatter={tooltipFormatter} labelFormatter={() => ''} />
-            <Legend formatter={legendFormatter} />
-            <Bar dataKey="count" fill="#8884d8" />
-          </BarChart>
-        </ResponsiveContainer>
-      )}
+      <PurchaseFrequencyChart fromDate={fromDate} toDate={toDate} />
     </div>
   )
 }
@@ -59,4 +32,4 @@ const PurchaseFrequencyTitle = styled.h3`
   margin-bottom: 20px;
 `
 
-export default PurchaseFrequencyChart
+export default PurchaseFrequency
